requirements.txt
# Snowflake Predictive Analytics Agent
# Production Dependencies

# Data Processing & Analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=12.0.0
joblib>=1.3.0

# Snowflake Connectivity
snowflake-connector-python>=3.5.0

# Web Framework & API
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
click>=8.1.0

# Caching & Storage
diskcache>=5.6.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
psutil>=5.9.0
